ci: add workaround for broken sudo in container
sudo seems to be broken within the container. This seems to be a known
issue, see for example [1].

The error is:
sudo: PAM account management error: Authentication service cannot retrieve authentication info
sudo: a password is required

This seems to be caused by 000 permissions on /etc/shadow, and can be
fixed with a simple chmod 0400 on /etc/shadow.

[1] rocky-linux/sig-cloud-instance-images#56

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>